requirements.txt
langchain
langchain_community
python-dotenv
openai
pypdf
docarray
tiktoken
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn[standard]>=0.15.0,<0.16.0